Include SPI Files, refactor a lot, do SPI and bootloader logic next

+++ src/microcontroller/va41620/microcontroller_dependent.cpp
@@ -6,9 +6,12 @@ void Move_Vector_Table()
 {
     constexpr uint32_t TBLOFF = 0xE000ED08;
     /* Set the vector table base address */
+    /** THis worked!*/
     uint32_t* vector_table = (uint32_t*)&__approm_start__; // NOLINT
-    auto* vtor = reinterpret_cast<uint32_t*>(TBLOFF); // NO LINT
-    *vtor = ((uint32_t)vector_table & 0xFFFFFFF8); // NOLINT
+    auto* vtor = reinterpret_cast<uint32_t*>(TBLOFF); // NOLINT
+    *vtor = ((uintptr_t)vector_table & 0xFFFFFFF8); // NOLINT
+    // TODO If anything fails, try this:
+    // *vtor = ((uint32_t)vector_table & 0xFFFFFFF8); // NOLINT
 }
 
 void Toggle_ROM_Writeable(bool writeable)
